+++ features/admin.applications.v1/hooks/use-async-operation-status.ts
@@ -24,27 +24,50 @@ import { MutableRefObject, useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useDispatch } from "react-redux";
 import { Dispatch } from "redux";
+import { AsyncOperationStatusResponse, OperationStatus, OperationStatusSummary } from
+    "../../admin.core.v1/models/common";
 import { getAsyncOperationStatus } from "../api/application";
-import { OperationStatus } from "../constants/application-management";
-import { OperationStatusSummary } from "../models/application";
 
 /**
  * Proptypes for the use operation status poller hook.
  */
 export interface useAsyncOperationStatusProps {
+    /**
+     * Type of the operation being tracked
+     */
     operationType: string;
+    /**
+     * The unique identifier of the subject associated with the operation.
+     */
     subjectId: string;
+    /**
+     * The interval between each poll to check the operation's status.
+     */
     pollingInterval?: number;
+    /**
+     * Callback function triggered when the operation reaches a completed status.
+     *
+     * @param finalStatus - The final status of the operation after completion.
+     */
     onCompleted?: (finalStatus: OperationStatus) => void;
-    onStatusChange?: (operationId: string, status: OperationStatus,
-        summary: OperationStatusSummary) => void;
+    /**
+     * Callback function triggered whenever the status of the operation changes.
+     *
+     * @param operationId - The unique identifier of the operation.
+     * @param status - The current status of the operation.
+     * @param summary - A summary of the operation's status.
+     */
+    onStatusChange?: (operationId: string, status: OperationStatus, summary: OperationStatusSummary) => void;
+    /**
+     * A flag to enable or disable the polling of the operation's status. If set to `false`, polling is disabled.
+     */
     enabled?: boolean;
 }
 
 /**
  * Proptypes for the REST api params.
  */
-interface Params {
+interface UseAsyncOperationStatusParams {
     filter?: string;
     limit?: number;
     after?: string;
@@ -52,11 +75,15 @@ interface Params {
 }
 
 /**
- *  Use operation status poller hook.
+ *  This hook polls the status of an asynchronous operation at a specified interval and triggers
+ *  the relevant callbacks when the status changes or the operation completes.
  *
- * @param props - Props injected to the component.
- *
- * @returns Use operation status poller hook.
+ * @param string - operationType - The type of the operation being tracked.
+ * @param string - subjectId - The unique identifier of the subject associated with the operation
+ * @param number - The interval between each poll to check the operation's status.
+ * @param Function - [onCompleted] - A callback function triggered when the operation reaches a completed status.
+ * @param Function - [onStatusChange] - A callback function triggered whenever the status of the operation changes.
+ * @param boolean - [enabled=true] - A flag to enable or disable the polling of the operation's status.
  */
 export const useAsyncOperationStatus = ({
     operationType,
@@ -67,34 +94,41 @@ export const useAsyncOperationStatus = ({
     enabled
 }: useAsyncOperationStatusProps) => {
 
-    const [ status, setStatus ] = useState<OperationStatus>(OperationStatus.IDLE);
-    const intervalRef: MutableRefObject<number> = useRef<number | null>(null);
-    const dispatch: Dispatch = useDispatch();
-    const { t } = useTranslation();
     const API_LIMIT: number = UIConstants.DEFAULT_RESOURCE_LIST_ITEM_LIMIT;
 
-    const [ params ] = useState<Params>({
+    const intervalRef: MutableRefObject<number> = useRef<number | null>(null);
+    const [ status, setStatus ] = useState<OperationStatus>(OperationStatus.IDLE);
+    const [ params ] = useState<UseAsyncOperationStatusParams>({
         after: null,
         filter: `subjectId eq ${subjectId} and operationType eq ${operationType}`,
         limit: API_LIMIT
     });
 
+    const dispatch: Dispatch = useDispatch();
+    const { t } = useTranslation();
+
+    /**
+     * Fetches the current status of the asynchronous operation.
+     */
     const fetchStatus = async (): Promise< { operationId: string, status: OperationStatus,
         summary: OperationStatusSummary } > => {
-
         const response: AxiosResponse = await getAsyncOperationStatus(params.after, null, params.filter, params.limit);
-        const newStatus: OperationStatus = response.data?.operations[0]?.status ?? OperationStatus.IDLE;
-        const newOperationId: string = response.data?.operations[0]?.operationId ?? null;
+        const operation: AsyncOperationStatusResponse = response.data?.operations[0];
+        const newStatus: OperationStatus = operation?.status ?? OperationStatus.IDLE;
+        const newOperationId: string = operation?.operationId ?? null;
         const operationSummary: OperationStatusSummary = {
-            failedCount: response.data?.operations[0]?.unitOperationDetail?.summary?.failed ?? 0,
-            partiallyCompletedCount: response.data?.operations[0]?.unitOperationDetail?.
-                summary?.partiallyCompleted ?? 0,
-            successCount: response.data?.operations[0]?.unitOperationDetail?.summary?.success ?? 0
+            failedCount: operation?.unitOperationDetail?.summary?.failed ?? 0,
+            partiallyCompletedCount: operation?.unitOperationDetail?.summary?.partiallyCompleted ?? 0,
+            successCount: operation?.unitOperationDetail?.summary?.success ?? 0
         };
 
         return { operationId: newOperationId, status: newStatus, summary: operationSummary };
     };
 
+    /**
+     * Starts polling for the operation status at regular intervals.
+     * Stops polling when the operation is completed.
+     */
     const startPolling = () => {
         if (!enabled || intervalRef.current !== null) return;
 
@@ -125,6 +159,10 @@ export const useAsyncOperationStatus = ({
         }, pollingInterval);
     };
 
+    /**
+     * Fetches the initial status of the asynchronous operation and starts polling if the status is in progress.
+     * Calls the `onCompleted` callback when the operation completes.
+     */
     const fetchInitialStatus = async () => {
         try {
