+++ features/admin.applications.v1/models/application.ts
@@ -939,61 +939,3 @@ export interface ShareApplicationStatusResponse {
     status: string;
     message: string;
 }
-
-/**
- * Interface for the share application status summary.
- */
-export interface OperationStatusSummary {
-    successCount: number;
-    failedCount: number;
-    partiallyCompletedCount: number
-}
-
-export interface AsyncOperationStatusUnitResponse {
-    unitOperationId: string;
-    operationId: string;
-    residentResourceId: string;
-    targetOrgId: string;
-    targetOrgName: string;
-    status: string;
-    statusMessage: string;
-    createdTime: string;
-}
-
-export interface AsyncOperationStatusResponse {
-    operationId: string;
-    correlationId: string;
-    operationType: string;
-    subjectType: string;
-    subjectId: string;
-    initiatedOrgId: string;
-    initiatedUserId: string;
-    status: string;
-    policy: string;
-    createdTime: string;
-    modifiedTime: string;
-    unitOperationDetail: {
-        ref: string;
-        summary: {
-            success: number;
-            failed: number;
-            partiallyCompleted: number;
-        }
-    }
-}
-
-export interface AsyncOperationStatusLinkInterface {
-    href: string;
-    rel: string;
-}
-
-export interface AsyncOperationStatusUnitListInterface {
-    links: AsyncOperationStatusLinkInterface[];
-    unitOperations: AsyncOperationStatusUnitResponse[];
-}
-
-export interface AsyncOperationStatusListInterface {
-    links: AsyncOperationStatusLinkInterface[];
-    operations: AsyncOperationStatusResponse[];
-}
-
